[JENKINS-51778] - Refresh the external logging demo
