--- demo/external-logging-elasticsearch/run.sh
@@ -1,2 +1,0 @@
-JENKINS_HOME=$(pwd)/work java -jar tmp/output/target/external-task-logging-logstash-1.0-SNAPSHOT.war --httpPort=8080 --prefix=/jenkins
-
