+++ demo/external-logging-elasticsearch/src/main/groovy/4_SaveToDisk.groovy
@@ -1,0 +1,3 @@
+import jenkins.model.Jenkins
+
+Jenkins.instance.save()
